Fix Interpolator

+++ src/main/java/frc/lib/Ranging/CargoBallInterpolator.java
@@ -37,8 +37,8 @@ public class CargoBallInterpolator {
      * @param distance the distance value.
      * @return the desired value for at the distance.
      */
-    public int calcMainShooterSpeed(double distance) {
-        int tempMainSpeed = 0;
+    public double calcMainShooterSpeed(double distance) {
+        double tempMainSpeed = 0;
         System.out.println("calcmainshooterspeed distance " + distance);
         if (dataPointList.size() == 1) {
             tempMainSpeed = dataPointList.get(0).getMainShooterSpeed();
@@ -61,8 +61,8 @@ public class CargoBallInterpolator {
             } else if (upperDataPoint.getDistance() == -1.0) {
                 tempMainSpeed = lowerDataPoint.getMainShooterSpeed();
             } else {
-                int upperMainSpeed = upperDataPoint.getMainShooterSpeed();
-                int lowerMainSpeed = lowerDataPoint.getMainShooterSpeed();
+                double upperMainSpeed = upperDataPoint.getMainShooterSpeed();
+                double lowerMainSpeed = lowerDataPoint.getMainShooterSpeed();
 
                 tempMainSpeed = lerp(lowerMainSpeed, upperMainSpeed, (distance - lowerDataPoint.getDistance())
                         / (upperDataPoint.getDistance() - lowerDataPoint.getDistance()));
@@ -75,8 +75,8 @@ public class CargoBallInterpolator {
         return tempMainSpeed;
     }
 
-    public int calcSecondShooterSpeed(double distance) {
-        int tempSecondSpeed = 0;
+    public double calcSecondShooterSpeed(double distance) {
+        double tempSecondSpeed = 0;
 
         if (dataPointList.size() == 1) {
             tempSecondSpeed = dataPointList.get(0).getSecondShooterSpeed();
@@ -98,8 +98,8 @@ public class CargoBallInterpolator {
             } else if (upperDataPoint.getDistance() == -1.0) {
                 tempSecondSpeed = lowerDataPoint.getSecondShooterSpeed();
             } else {
-                int upperSecondSpeed = upperDataPoint.getSecondShooterSpeed();
-                int lowerSecondSpeed = lowerDataPoint.getSecondShooterSpeed();
+                double upperSecondSpeed = upperDataPoint.getSecondShooterSpeed();
+                double lowerSecondSpeed = lowerDataPoint.getSecondShooterSpeed();
 
                 tempSecondSpeed = lerp(lowerSecondSpeed, upperSecondSpeed, (distance - lowerDataPoint.getDistance())
                         / (upperDataPoint.getDistance() - lowerDataPoint.getDistance()));
